requirements.txt
streamlit
openai
firecrawl-py
pandas
plotly
python-dotenv
requests